rename files...
...and rearrange activity-handling
